Ohare.py bootstraps EDA
Project is set up technically as a Flask app. But for now, I have wrote a Class to bootstrap EDA, and assist with doing exploratory work in the notebooks/ directory. The data is too large to upload to GH, so references to data/ will not work... FAA BTA has files.